requirements.txt
streamlit
PyPDF2
pandas
google-genai
